Split SQLite connection helpers
