+++ internal/db/migrate_legacy.go
@@ -5,169 +5,10 @@ import (
 	"database/sql"
 	"encoding/json"
 	"fmt"
-	"net/url"
-	"os"
-	"path/filepath"
 	"sort"
 	"strconv"
 	"strings"
-
-	_ "github.com/mattn/go-sqlite3" // sqlite driver
-	"github.com/rs/zerolog/log"
 )
-
-type DB struct {
-	*sql.DB
-	*Queries
-}
-
-// Open initializes the SQLite database connection with high performance pragmas and runs schema migrations.
-func Open(databaseURL string) (*DB, error) {
-	log.Info().Str("url", databaseURL).Msg("Connecting to SQLite database")
-
-	dsn, err := normalizeSQLiteDSN(databaseURL)
-	if err != nil {
-		return nil, err
-	}
-	if err := ensureSQLiteDir(dsn); err != nil {
-		return nil, err
-	}
-
-	db, err := sql.Open("sqlite3", dsn)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open database: %w", err)
-	}
-	db.SetMaxOpenConns(8)
-	db.SetMaxIdleConns(8)
-	db.SetConnMaxLifetime(0)
-
-	if err := db.Ping(); err != nil {
-		return nil, fmt.Errorf("failed to ping database: %w", err)
-	}
-
-	if _, err := db.ExecContext(context.Background(), `
-		PRAGMA journal_mode = WAL;
-		PRAGMA synchronous = NORMAL;
-		PRAGMA busy_timeout = 10000;
-		PRAGMA foreign_keys = ON;
-		PRAGMA wal_autocheckpoint = 1000;
-	`); err != nil {
-		return nil, fmt.Errorf("failed to configure SQLite pragmas: %w", err)
-	}
-
-	instance := &DB{
-		DB:      db,
-		Queries: New(db),
-	}
-
-	if err := instance.migrate(); err != nil {
-		return nil, fmt.Errorf("failed to run migrations: %w", err)
-	}
-
-	// Enforce pragmas for any connections opened after migration.
-	if _, err := db.ExecContext(context.Background(), `
-		PRAGMA journal_mode = WAL;
-		PRAGMA synchronous = NORMAL;
-		PRAGMA busy_timeout = 10000;
-		PRAGMA foreign_keys = ON;
-		PRAGMA wal_autocheckpoint = 1000;
-	`); err != nil {
-		log.Warn().Err(err).Msg("Failed to enforce some SQLite pragmas")
-	}
-	if err := hardenSQLiteFiles(dsn); err != nil {
-		_ = db.Close()
-		return nil, err
-	}
-
-	log.Info().Msg("Database connected and configured successfully")
-	return instance, nil
-}
-
-func normalizeSQLiteDSN(databaseURL string) (string, error) {
-	databaseURL = strings.TrimSpace(databaseURL)
-	if databaseURL == "" {
-		databaseURL = "file:p2pstream.db?mode=rwc"
-	}
-
-	if strings.HasPrefix(databaseURL, "file:") {
-		prefix, rawQuery, _ := strings.Cut(databaseURL, "?")
-		values, err := url.ParseQuery(rawQuery)
-		if err != nil {
-			return "", fmt.Errorf("invalid sqlite database URL %q: %w", databaseURL, err)
-		}
-		if values.Get("mode") == "" && prefix != "file::memory:" {
-			values.Set("mode", "rwc")
-		}
-		applySQLitePragmas(values)
-		return prefix + "?" + values.Encode(), nil
-	}
-
-	values := url.Values{}
-	values.Set("mode", "rwc")
-	applySQLitePragmas(values)
-	return "file:" + databaseURL + "?" + values.Encode(), nil
-}
-
-func applySQLitePragmas(values url.Values) {
-	values.Set("_journal_mode", "WAL")
-	values.Set("_synchronous", "NORMAL")
-	values.Set("_busy_timeout", "10000")
-	values.Set("_fk", "1")
-	values.Set("cache", "private")
-}
-
-func ensureSQLiteDir(dsn string) error {
-	path, ok := sqliteFilePathFromDSN(dsn)
-	if !ok {
-		return nil
-	}
-	dir := filepath.Dir(path)
-	if dir == "." || dir == "" {
-		return nil
-	}
-	if err := os.MkdirAll(dir, 0700); err != nil {
-		return err
-	}
-	return os.Chmod(dir, 0700)
-}
-
-func hardenSQLiteFiles(dsn string) error {
-	path, ok := sqliteFilePathFromDSN(dsn)
-	if !ok {
-		return nil
-	}
-	for _, candidate := range []string{path, path + "-wal", path + "-shm"} {
-		if _, err := os.Stat(candidate); err != nil {
-			if os.IsNotExist(err) {
-				continue
-			}
-			return fmt.Errorf("failed to stat sqlite file %q: %w", candidate, err)
-		}
-		if err := os.Chmod(candidate, 0600); err != nil {
-			return fmt.Errorf("failed to secure sqlite file %q: %w", candidate, err)
-		}
-	}
-	return nil
-}
-
-func sqliteFilePathFromDSN(dsn string) (string, bool) {
-	if !strings.HasPrefix(dsn, "file:") {
-		return "", false
-	}
-	prefix, rawQuery, _ := strings.Cut(dsn, "?")
-	values, err := url.ParseQuery(rawQuery)
-	if err == nil && strings.EqualFold(values.Get("mode"), "memory") {
-		return "", false
-	}
-	path := strings.TrimPrefix(prefix, "file:")
-	if path == "" || path == ":memory:" || strings.HasPrefix(path, ":memory:") {
-		return "", false
-	}
-	if unescaped, err := url.PathUnescape(path); err == nil {
-		path = unescaped
-	}
-	return filepath.Clean(path), true
-}
 
 // migrate runs the initial schema setup.
 func (db *DB) migrate() error {
